made WorldGenerators not be hardcoded at all anymore, welcome addons!
