+++ common/src/main/java/com/gregtechceu/gtceu/api/data/worldgen/IWorldGenLayer.java
@@ -6,6 +6,11 @@ import net.minecraft.util.ExtraCodecs;
 import net.minecraft.util.StringRepresentable;
 import net.minecraft.world.level.levelgen.structure.templatesystem.RuleTest;
 
+/**
+ * @author Screret
+ * @date 2023/6/9
+ * @implNote IWorldGenLayer
+ */
 public interface IWorldGenLayer extends StringRepresentable {
     Codec<IWorldGenLayer> CODEC = ExtraCodecs.stringResolverCodec(StringRepresentable::getSerializedName, WorldGeneratorUtils.WORLD_GEN_LAYERS::get);
