+++ common/src/main/java/com/gregtechceu/gtceu/api/data/worldgen/WorldGenLayers.java
@@ -13,6 +13,11 @@ import net.minecraft.util.StringRepresentable;
 import net.minecraft.world.level.levelgen.structure.templatesystem.RuleTest;
 import org.jetbrains.annotations.NotNull;
 
+/**
+ * @author Screret
+ * @date 2023/6/9
+ * @implNote WorldGenLayers
+ */
 public enum WorldGenLayers implements IWorldGenLayer, StringRepresentable {
     STONE("stone", OreFeatures.STONE_ORE_REPLACEABLES),
     DEEPSLATE("deepslate", OreFeatures.DEEPSLATE_ORE_REPLACEABLES),
@@ -20,6 +25,7 @@ public enum WorldGenLayers implements IWorldGenLayer, StringRepresentable {
     ENDSTONE("endstone", GTOres.END_ORE_REPLACEABLES);
 
     private final String name;
+    @SuppressWarnings("NonFinalFieldInEnum")
     @Getter @Setter
     private RuleTest target;
